New asynchronous IPC primitive: pulses
Each handle has a bitmask of up to 64 events that can be signaled by
setting one or more bits. Sending a pulse never blocks, so it is safe to
use for a server that wants to notify one of its clients about finished
work or other events.

In a receive that can receive from a given handle, all pending pulses will
be received first, and the act of receiving them will implicitly clear the
mask of signaled/pending pulses.