e1000: Rewrite :)
First, use the pulse API to send notifications about incoming packets
and finished sends to fix the deadlock problems.

The second change is a completely asynchronous API, and dynamic buffer
assignment allowing protocols to have several buffers (and control
allocation of send vs receive buffers) as well as allowing the ethernet
driver's buffer allocation to be independent from the protocol's.