Implement a kernel test suite (one test so far)
Based on a DSL in Python, generate code for a number of communicating
processes and a master process that sequences and validates output.
Compile the test code to binaries and launch them with qemu.

The master program writes PASS or FAIL on the qemu debug console, and
the Python testing code parses the output looking for panic, pass or
fail.

Adds a `make test` target that builds the necessary objects and kernel
and runs on both the assembly and C++ kernels.